Add webp decoder.
We've been needing this for a while for the game engines but we also need it to support decoding webp images for use in the editor with the Rive Renderer.

I haven't instrumented the build properly to use SIMD extensions, but I left some notes for how to do so. This PR unblocks the use of WebP, let's do some perf improvements in a follow up that perhaps the runtime team can own?

Diffs=
160d9eefb Add webp decoder. (#7883)