Nnnn text alignment and fit
added an explicit option for users to respect the baseline when text is vertical aligned.
The checkbox is disabled if the overflow options is other than Fit.
When the baseline is not accounted for, text can grow larger in the text area.

https://github.com/user-attachments/assets/e43b3dbf-3ca4-4e4e-8d75-8beb401108d6

Diffs=
b3534f1f8f Nnnn text alignment and fit (#8385)