Vulkan - Atlas texture barrier fix
The barrier that occurs after the atlas texture render pass was incorrectly performing the barrier on the tessellation vertex texture. Curious how this will look on browserstack, because this barrier could cause corrupted data in the tess texture. If there is no atlas, the last texture access to the atlas in the VK_UNDEFINED layout, which is a "don't care" for the existing data. The rendering code is fragile in its current state!

Diffs=
48b57c6062 Vulkan - Atlas texture barrier fix  (#9212)